Chat and UI overhaul

+++ src/components/construct-box/index.tsx
@@ -12,8 +12,9 @@ import { saveTavernCardAsImage } from "@/api/extrasapi";
 import { Download } from "lucide-react";
 interface Props {
     character: Construct;
+    onCharacterDelete: (character: Construct) => void;
 }
-const ConstructBox: React.FC<Props> = ({character}) => {
+const ConstructBox: React.FC<Props> = ({character, onCharacterDelete}) => {
     const [characterName, setCharacterName] = useState<string>(character.name);
     const [isOpen, setIsOpen] = useState<boolean>(false);
     const [isActive, setIsActive] = useState<boolean>(false);
@@ -49,23 +50,27 @@ const ConstructBox: React.FC<Props> = ({character}) => {
     }, [character]);
 
     const deleteConstructFrom = async () => {
+        if(!confirm(`Are you sure you want to delete this construct? This cannot be undone.`)) return;
         await deleteConstruct(character._id);
-        window.location.reload();
+        onCharacterDelete(character);
     }
 
     const makeActive = async () => {
         await addConstructToActive(character._id);
-        window.location.reload();
+        setIsActive(true);
+        setIsPrimary(false);
     }
 
     const makePrimary = async () => {
         await setConstructAsPrimary(character._id);
-        window.location.reload();
+        setIsPrimary(true);
+        setIsActive(true);
     }
 
     const makeInactive = async () => {
         await removeConstructFromActive(character._id);
-        window.location.reload();
+        setIsActive(false);
+        setIsPrimary(false);
     }
 
     useEffect(() => {
